fix: conflict symbols issue with mkfs
