+++ inc/file.h
@@ -48,6 +48,8 @@ struct devsw {
 
 extern struct devsw devsw[];
 
+// kern/file.c
+
 void fileinit();
 struct file* filealloc();
 struct file* filedup(struct file*);
@@ -55,5 +57,28 @@ void fileclose(struct file*);
 int filestat(struct file*, struct stat*);
 ssize_t fileread(struct file*, char*, ssize_t);
 ssize_t filewrite(struct file*, char*, ssize_t);
+
+// kern/fs.c
+
+void readsb(int, struct superblock*);
+
+void iinit(int);
+struct inode* ialloc(uint32_t, uint16_t);
+void iupdate(struct inode*);
+struct inode* idup(struct inode*);
+void ilock(struct inode*);
+void iunlock(struct inode*);
+void iput(struct inode*);
+void iunlockput(struct inode*);
+void stati(struct inode*, struct stat*);
+ssize_t readi(struct inode*, char*, size_t, size_t);
+ssize_t writei(struct inode*, char*, size_t, size_t);
+
+int namecmp(const char*, const char*);
+struct inode* dirlookup(struct inode*, char*, size_t*);
+int dirlink(struct inode*, char*, uint32_t);
+
+struct inode* namei(char*);
+struct inode* nameiparent(char*, char*);
 
 #endif  // INC_FILE_H_
