+++ kern/fs.c
@@ -12,6 +12,7 @@
  */
 
 #include "fs.h"
+
 #include "buf.h"
 #include "console.h"
 #include "file.h"
